@@ -13,9 +13,12 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.oauth2.server.resource.OAuth2ResourceServerConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.transaction.PlatformTransactionManager;
 
 @SpringBootApplication
@@ -34,21 +37,27 @@ public class ServicesCardApiApplication {
    * Add security exceptions for swagger UI and prometheus.
    */
   @Configuration
+  @EnableMethodSecurity
   static
-  class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
+  class WebSecurityConfiguration{
 
-    @Override
-    public void configure(WebSecurity web) {
-      web.ignoring().antMatchers("/v3/api-docs/**",
-              "/actuator/health","/actuator/prometheus",
-              "/swagger-ui/**", "/health");
+    public WebSecurityConfiguration() {
+      super();
     }
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
       http
-              .authorizeRequests()
-              .anyRequest().authenticated().and()
-              .oauth2ResourceServer().jwt();
+          .csrf(AbstractHttpConfigurer::disable)
+          .authorizeHttpRequests(auth -> auth
+              .requestMatchers("/v3/api-docs/**",
+                  "/actuator/health", "/actuator/prometheus","/actuator/**",
+                  "/swagger-ui/**").permitAll()
+              .anyRequest().authenticated()
+          )
+          .sessionManagement(sess -> sess.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+          .oauth2ResourceServer(OAuth2ResourceServerConfigurer::jwt);
+      return http.build();
     }
   }
   @Bean
