@@ -7,8 +7,8 @@ import lombok.val;
 import org.slf4j.MDC;
 import org.springframework.lang.NonNull;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
